chore(refactor): merged addCredentialCommand + addTargetCommand
