+++ src/commands/add/addCommand.ts
@@ -2,26 +2,51 @@ import { TargetScope } from '../../types'
 import { CommandExample, ReturnCode } from '../../types/command'
 import { TargetCommand } from '../../types/command/targetCommand'
 import { addCredential } from './addCredential'
+import { addTarget } from './addTarget'
+
+enum AddSubCommand {
+  Cred = 'cred'
+}
+
+enum AddAlias {
+  Auth = 'auth'
+}
 
 // The syntax here needs to use square brackets for the subCommand(i.e. make it optional) since
 // the alias `auth` does not have a subCommand
-const syntax = 'add [subCommand]'
-const aliases = ['auth']
-const usage = 'sasjs add cred [options] | sasjs auth [options]'
+const syntax = 'add [subCommand] [options]'
+const aliases = Object.values(AddAlias)
+const usage = 'sasjs add | sasjs add cred [options] | sasjs auth [options]'
 const description =
   'Authenticates against the specified target and creates a `.env.{targetName}` file at the root of the current project.'
 const examples: CommandExample[] = [
   {
-    command: 'sasjs add cred -t myTarget',
+    command: 'sasjs add',
     description: ''
   },
   {
-    command: 'sasjs auth -t myTarget',
+    command: 'sasjs add --insecure',
+    description: ''
+  },
+  {
+    command: 'sasjs add cred --target <target-name> --insecure',
+    description: ''
+  },
+  {
+    command: 'sasjs add cred -t <target-name> -i',
+    description: ''
+  },
+  {
+    command: 'sasjs auth --target <target-name> --insecure',
+    description: ''
+  },
+  {
+    command: 'sasjs auth -t <target-name> -i',
     description: ''
   }
 ]
 
-export class AddCredentialCommand extends TargetCommand {
+export class AddCommand extends TargetCommand {
   constructor(args: string[]) {
     const parseOptions: { [key: string]: Object } = {
       insecure: {
@@ -40,6 +65,15 @@ export class AddCredentialCommand extends TargetCommand {
   }
 
   public async execute() {
+    const commandNameEntered = this.parsed._[0]
+
+    return commandNameEntered === AddAlias.Auth ||
+      this.parsed.subCommand === AddSubCommand.Cred
+      ? await this.executeCred()
+      : await this.executeTarget()
+  }
+
+  public async executeCred() {
     const { target, isLocal } = await this.getTargetInfo()
     const scope = isLocal ? TargetScope.Local : TargetScope.Global
     return await addCredential(target, this.insecure, scope)
@@ -49,6 +83,18 @@ export class AddCredentialCommand extends TargetCommand {
       })
       .catch((err) => {
         process.logger?.error('Error adding credentials: ', err)
+        return ReturnCode.InternalError
+      })
+  }
+
+  public async executeTarget() {
+    return await addTarget(this.insecure)
+      .then(() => {
+        process.logger?.success('Target has been successfully added!')
+        return ReturnCode.Success
+      })
+      .catch((err) => {
+        process.logger?.error('Error adding target: ', err)
         return ReturnCode.InternalError
       })
   }
